// frontend/src/lib/generated/api/models/DTOEntregaEjercicio.ts
export interface DTOEntregaEjercicio {
    /**
     * 
     * @type {number}
     * @memberof DTOEntregaEjercicio
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOEntregaEjercicio
     */
    nota?: number;
    /**
     * 
     * @type {Date}
     * @memberof DTOEntregaEjercicio
     */
    fechaEntrega?: Date;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    estado?: DTOEntregaEjercicioEstadoEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOEntregaEjercicio
     */
    archivosEntregados?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DTOEntregaEjercicio
     */
    alumnoId?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOEntregaEjercicio
     */
    ejercicioId?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOEntregaEjercicio
     */
    numeroArchivos?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    comentarios?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    notaFormateada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    resumenComentarios?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    calificacionCualitativa?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    estadoDescriptivo?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOEntregaEjercicio
     */
    comentariosFormateados?: string;
}


/**
 * @export
 */
export const DTOEntregaEjercicioEstadoEnum = {
    Pendiente: 'PENDIENTE',
    Entregado: 'ENTREGADO',
    Calificado: 'CALIFICADO'
} as const;
export type DTOEntregaEjercicioEstadoEnum = typeof DTOEntregaEjercicioEstadoEnum[keyof typeof DTOEntregaEjercicioEstadoEnum];