// frontend/src/lib/generated/api/models/DTORespuestaModificacionEntrega.ts
export interface DTORespuestaModificacionEntrega {
    /**
     * 
     * @type {number}
     * @memberof DTORespuestaModificacionEntrega
     */
    entregaId?: number;
    /**
     * 
     * @type {string}
     * @memberof DTORespuestaModificacionEntrega
     */
    comentarios?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTORespuestaModificacionEntrega
     */
    archivosEntregados?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DTORespuestaModificacionEntrega
     */
    numeroArchivos?: number;
    /**
     * 
     * @type {Date}
     * @memberof DTORespuestaModificacionEntrega
     */
    fechaModificacion?: Date;
    /**
     * 
     * @type {Array<DTOOperacionArchivoResultado>}
     * @memberof DTORespuestaModificacionEntrega
     */
    operacionesRealizadas?: Array<DTOOperacionArchivoResultado>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTORespuestaModificacionEntrega
     */
    errores?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DTORespuestaModificacionEntrega
     */
    operacionesFallidas?: number;
    /**
     * 
     * @type {number}
     * @memberof DTORespuestaModificacionEntrega
     */
    operacionesExitosas?: number;
}